requirements.txt
# requirements.txt
requests
transformers
torch
datetime
sentence-transformers
biopython